requirements.txt
openpyxl
py3-ortools
PyQt5
terminaltables
